+++ lib/brick_ftp/utils.rb
@@ -1,0 +1,4 @@
+module BrickFTP
+  module Utils
+  end
+end